requirements.txt
numpy==1.18.2
pathlib2==2.3.5
torch==1.11.0+cu113
torchvision==0.12.0+cu113
opencv-python==4.5.5.62
Pillow==7.2.0
mediapipe==0.8.9.1
matplotlib==3.5.1